Fix warnings

+++ chess/position.cpp
@@ -143,11 +143,11 @@ bool position::verify( std::string& error ) const
 		error = "Black king count is not exactly one";
 		return false;
 	}
-	if( bitscan( bitboards[color::white].b[bb_type::king]) != king_pos[color::white] ) {
+	if( static_cast<int>(bitscan( bitboards[color::white].b[bb_type::king])) != king_pos[color::white] ) {
 		error = "Internal error: White king position does not match king bitboard";
 		return false;
 	}
-	if( bitscan( bitboards[color::black].b[bb_type::king]) != king_pos[color::black] ) {
+	if( static_cast<int>(bitscan( bitboards[color::black].b[bb_type::king])) != king_pos[color::black] ) {
 		error = "Internal error: Black king position does not match king bitboard";
 		return false;
 	}
